Move isActionOfType to reduxHelper
