+++ visualization/app/codeCharta/state/store/appSettings/dynamicMargin/dynamicMargin.service.ts
@@ -1,7 +1,7 @@
 import { StoreService, StoreSubscriber } from "../../../store.service"
 import { IRootScopeService } from "angular"
 import { DynamicMarginActions } from "./dynamicMargin.actions"
-import { isActionOfType } from "../../../../util/actionHelper"
+import { isActionOfType } from "../../../../util/reduxHelper"
 
 export interface DynamicMarginSubscriber {
 	onDynamicMarginChanged(dynamicMargin: boolean)
