+++ visualization/app/codeCharta/state/store/appSettings/isPresentationMode/isPresentationMode.service.ts
@@ -1,7 +1,7 @@
 import { StoreService, StoreSubscriber } from "../../../store.service"
 import { IRootScopeService } from "angular"
 import { PresentationModeActions } from "./isPresentationMode.actions"
-import { isActionOfType } from "../../../../util/actionHelper"
+import { isActionOfType } from "../../../../util/reduxHelper"
 
 export interface IsPresentationModeSubscriber {
 	onPresentationModeChanged(isPresentationMode: boolean)
