+++ visualization/app/codeCharta/state/store/appSettings/isWhiteBackground/isWhiteBackground.service.ts
@@ -1,7 +1,7 @@
 import { StoreService, StoreSubscriber } from "../../../store.service"
 import { IRootScopeService } from "angular"
 import { IsWhiteBackgroundActions } from "./isWhiteBackground.actions"
-import { isActionOfType } from "../../../../util/actionHelper"
+import { isActionOfType } from "../../../../util/reduxHelper"
 
 export interface IsWhiteBackgroundSubscriber {
 	onIsWhiteBackgroundChanged(isWhiteBackground: boolean)
