+++ visualization/app/codeCharta/state/store/dynamicSettings/areaMetric/areaMetric.service.ts
@@ -4,7 +4,7 @@ import { AreaMetricActions, setAreaMetric } from "./areaMetric.actions"
 import { MetricData } from "../../../../codeCharta.model"
 import { MetricService, MetricServiceSubscriber } from "../../../metric.service"
 import { getMetricNameFromIndexOrLast, isAnyMetricAvailable, isMetricUnavailable } from "../../../../util/metricHelper"
-import { isActionOfType } from "../../../../util/actionHelper"
+import { isActionOfType } from "../../../../util/reduxHelper"
 
 export interface AreaMetricSubscriber {
 	onAreaMetricChanged(areaMetric: string)
