+++ visualization/app/codeCharta/state/store/dynamicSettings/distributionMetric/distributionMetric.service.ts
@@ -4,7 +4,7 @@ import { DistributionMetricActions, setDistributionMetric } from "./distribution
 import { MetricData } from "../../../../codeCharta.model"
 import { MetricService, MetricServiceSubscriber } from "../../../metric.service"
 import { getMetricNameFromIndexOrLast, isAnyMetricAvailable, isMetricUnavailable } from "../../../../util/metricHelper"
-import { isActionOfType } from "../../../../util/actionHelper"
+import { isActionOfType } from "../../../../util/reduxHelper"
 
 export interface DistributionMetricSubscriber {
 	onDistributionMetricChanged(distributionMetric: string)
