+++ visualization/app/codeCharta/state/store/dynamicSettings/searchPattern/searchPattern.service.ts
@@ -3,7 +3,7 @@ import { IRootScopeService } from "angular"
 import { SearchPatternActions, setSearchPattern } from "./searchPattern.actions"
 import { FilesService, FilesSelectionSubscriber } from "../../files/files.service"
 import { Files } from "../../../../model/files"
-import { isActionOfType } from "../../../../util/actionHelper"
+import { isActionOfType } from "../../../../util/reduxHelper"
 
 export interface SearchPatternSubscriber {
 	onSearchPatternChanged(searchPattern: string)
