+++ visualization/app/codeCharta/state/store/treeMap/mapSize/mapSize.service.ts
@@ -1,7 +1,7 @@
 import { StoreService, StoreSubscriber } from "../../../store.service"
 import { IRootScopeService } from "angular"
 import { MapSizeActions } from "./mapSize.actions"
-import { isActionOfType } from "../../../../util/actionHelper"
+import { isActionOfType } from "../../../../util/reduxHelper"
 
 export interface MapSizeSubscriber {
 	onMapSizeChanged(mapSize: number)
