+++ visualization/app/codeCharta/ui/codeMap/codeMap.preRender.service.ts
@@ -14,11 +14,8 @@ import { ScalingService, ScalingSubscriber } from "../../state/store/appSettings
 import _ from "lodash"
 import { ScalingActions } from "../../state/store/appSettings/scaling/scaling.actions"
 import { IsLoadingMapActions, setIsLoadingMap } from "../../state/store/appSettings/isLoadingMap/isLoadingMap.actions"
-import {
-	IsLoadingFileActions,
-	setIsLoadingFile
-} from "../../state/store/appSettings/isLoadingFile/isLoadingFile.actions"
-import { isActionOfType } from "../../util/actionHelper"
+import { IsLoadingFileActions, setIsLoadingFile } from "../../state/store/appSettings/isLoadingFile/isLoadingFile.actions"
+import { isActionOfType } from "../../util/reduxHelper"
 
 const clone = require("rfdc")()
