--- visualization/app/codeCharta/util/actionHelper.ts
@@ -1,3 +1,0 @@
-export function isActionOfType(actionType: string, actions) {
-	return Object.values(actions).includes(actionType)
-}
