+++ visualization/app/codeCharta/util/reduxHelper.spec.ts
@@ -1,4 +1,6 @@
-import { addItemToArray, removeItemFromArray } from "./reduxHelper"
+import { addItemToArray, isActionOfType, removeItemFromArray } from "./reduxHelper"
+import { ScalingActions } from "../state/store/appSettings/scaling/scaling.actions"
+import { IsLoadingFileActions } from "../state/store/appSettings/isLoadingFile/isLoadingFile.actions"
 
 describe("reduxHelper", () => {
 	let obj1: { x: number; y: number }
@@ -36,6 +38,20 @@ describe("reduxHelper", () => {
 			mutateObject(obj2)
 
 			expect(result).toEqual([{ x: 10000, y: 2 }, { x: 3, y: 4 }])
+		})
+	})
+
+	describe("isActionOfType", () => {
+		it("should return true if an action is part of an enum", () => {
+			const result = isActionOfType(ScalingActions.SET_SCALING, ScalingActions)
+
+			expect(result).toBeTruthy()
+		})
+
+		it("should return false if an action is not part of an enum", () => {
+			const result = isActionOfType(IsLoadingFileActions.SET_IS_LOADING_FILE, ScalingActions)
+
+			expect(result).toBeFalsy()
 		})
 	})
 })
